chore(e2e): split up loss-payee and your-buyer e2e tests
