+++ .github/workflows/test.yml
@@ -237,6 +237,7 @@ jobs:
             'check-your-answers/policy/change-your-answers/another-company/*.spec.js',
             'check-your-answers/policy/change-your-answers/broker/*.spec.js',
             'check-your-answers/policy/change-your-answers/loss-payee/*.spec.js',
+            'check-your-answers/policy/change-your-answers/loss-payee/financial-details/*.spec.js',
             'check-your-answers/policy/change-your-answers/multiple-policy/*.spec.js',
             'check-your-answers/policy/change-your-answers/name-on-policy/*.spec.js',
             'check-your-answers/policy/change-your-answers/other-company-details/*.spec.js',
@@ -245,7 +246,10 @@ jobs:
             'check-your-answers/policy/summary-list/multiple-policy/*.spec.js',
             'check-your-answers/policy/summary-list/single-policy/*.spec.js',
             'check-your-answers/your-business/**/*.spec.js',
-            'check-your-answers/your-buyer/**/*.spec.js',
+            'check-your-answers/your-buyer/*.spec.js',
+            'check-your-answers/your-buyer/connection-with-buyer/*.spec.js',
+            'check-your-answers/your-buyer/trading-history/*.spec.js',
+            'check-your-answers/your-buyer/working-with-buyer/*.spec.js',
             'cookies-consent/**/*.spec.js',
             'dashboard/**/*.spec.js',
             'declarations/anti-bribery/**/*.spec.js',
@@ -299,7 +303,8 @@ jobs:
             'policy/broker-confirm-address/**/*.spec.js',
             'policy/change-your-answers/another-company/**/*.spec.js',
             'policy/change-your-answers/broker/**/*.spec.js',
-            'policy/change-your-answers/loss-payee/**/*.spec.js',
+            'policy/change-your-answers/loss-payee/*.spec.js',
+            'policy/change-your-answers/loss-payee/financial-details/*.spec.js',
             'policy/change-your-answers/multiple-policy/**/*.spec.js',
             'policy/change-your-answers/name-on-policy/**/*.spec.js',
             'policy/change-your-answers/other-company-details/**/*.spec.js',
@@ -335,7 +340,10 @@ jobs:
             'your-business/turnover/**/*.spec.js',
             'your-buyer/alternative-currency/**/*.spec.js',
             'your-buyer/buyer-financial-information/**/*.spec.js',
-            'your-buyer/change-your-answers/**/*.spec.js',
+            'your-buyer/change-your-answers/*.spec.js',
+            'your-buyer/change-your-answers/connection-with-buyer/*.spec.js',
+            'your-buyer/change-your-answers/trading-history/*.spec.js',
+            'your-buyer/change-your-answers/working-with-buyer/*.spec.js',
             'your-buyer/check-your-answers/**/*.spec.js',
             'your-buyer/company-or-organisation/**/*.spec.js',
             'your-buyer/connection-with-the-buyer/**/*.spec.js',
