+++ e2e-tests/insurance/cypress/e2e/journeys/check-your-answers/policy/change-your-answers/loss-payee/financial-details/change-your-answers-change-loss-payee-financial-details-international-to-uk.spec.js
@@ -1,8 +1,8 @@
-import partials from '../../../../../../../../partials';
-import { field, summaryList } from '../../../../../../../../pages/shared';
-import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance/policy';
-import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
-import checkSummaryList from '../../../../../../../../commands/insurance/check-policy-summary-list';
+import partials from '../../../../../../../../../partials';
+import { field, summaryList } from '../../../../../../../../../pages/shared';
+import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../../../constants/field-ids/insurance/policy';
+import { INSURANCE_ROUTES } from '../../../../../../../../../constants/routes/insurance';
+import checkSummaryList from '../../../../../../../../../commands/insurance/check-policy-summary-list';
 
 const {
   LOSS_PAYEE_DETAILS: { NAME, LOCATION, IS_LOCATED_IN_UK, IS_LOCATED_INTERNATIONALLY },
