+++ e2e-tests/insurance/cypress/e2e/journeys/check-your-answers/policy/change-your-answers/loss-payee/financial-details/change-your-answers-change-loss-payee-financial-details-international.spec.js
@@ -1,11 +1,11 @@
-import partials from '../../../../../../../../partials';
-import { field, summaryList } from '../../../../../../../../pages/shared';
-import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance/policy';
-import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
-import { POLICY_FIELDS as FIELDS } from '../../../../../../../../content-strings/fields/insurance/policy';
-import { mockBicSwiftCode1 } from '../../../../../../../../fixtures/bic-swift-codes';
-import { mockIban1 } from '../../../../../../../../fixtures/ibans';
-import { mockAddress1 } from '../../../../../../../../fixtures/addresses';
+import partials from '../../../../../../../../../partials';
+import { field, summaryList } from '../../../../../../../../../pages/shared';
+import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../../../constants/field-ids/insurance/policy';
+import { INSURANCE_ROUTES } from '../../../../../../../../../constants/routes/insurance';
+import { POLICY_FIELDS as FIELDS } from '../../../../../../../../../content-strings/fields/insurance/policy';
+import { mockBicSwiftCode1 } from '../../../../../../../../../fixtures/bic-swift-codes';
+import { mockIban1 } from '../../../../../../../../../fixtures/ibans';
+import { mockAddress1 } from '../../../../../../../../../fixtures/addresses';
 
 const {
   LOSS_PAYEE_FINANCIAL_INTERNATIONAL: { BIC_SWIFT_CODE, IBAN },
