+++ e2e-tests/insurance/cypress/e2e/journeys/check-your-answers/your-buyer/change-your-answers/trading-history/change-your-answers-change-trading-history-outstanding-payments-yes-to-no.spec.js
@@ -1,21 +1,27 @@
-import { field, summaryList } from '../../../../../../pages/shared';
-import { FIELD_VALUES } from '../../../../../../constants';
-import { YOUR_BUYER as FIELD_IDS } from '../../../../../../constants/field-ids/insurance/your-buyer';
-import { INSURANCE_ROUTES } from '../../../../../../constants/routes/insurance';
+import { field, status, summaryList } from '../../../../../../../../pages/shared';
+import partials from '../../../../../../../../partials';
+import { FIELD_VALUES } from '../../../../../../../../constants';
+import { YOUR_BUYER as FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance/your-buyer';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
 
 const { OUTSTANDING_PAYMENTS, TOTAL_AMOUNT_OVERDUE, TOTAL_OUTSTANDING_PAYMENTS } = FIELD_IDS;
 
 const {
   ROOT,
-  YOUR_BUYER: { TRADING_HISTORY, TRADING_HISTORY_CHANGE, CHECK_YOUR_ANSWERS },
+  CHECK_YOUR_ANSWERS: { YOUR_BUYER },
+  YOUR_BUYER: { TRADING_HISTORY, TRADING_HISTORY_CHECK_AND_CHANGE },
 } = INSURANCE_ROUTES;
+
+const { taskList } = partials.insurancePartials;
+
+const task = taskList.submitApplication.tasks.checkAnswers;
 
 const fieldId = OUTSTANDING_PAYMENTS;
 
 const baseUrl = Cypress.config('baseUrl');
 
 context(
-  `Insurance - Your buyer - Change your answers - Trading history - ${OUTSTANDING_PAYMENTS} - Yes to no - As an exporter, I want to change my answers to the trading history section`,
+  `Insurance - Check your answers - Your buyer - Trading history - ${OUTSTANDING_PAYMENTS} - Yes to no - As an exporter, I want to change my answers to the trading history section`,
   () => {
     let referenceNumber;
     let url;
@@ -25,20 +31,28 @@ context(
       cy.completeSignInAndGoToApplication({}).then(({ referenceNumber: refNumber }) => {
         referenceNumber = refNumber;
 
-        cy.startInsuranceYourBuyerSection({});
+        cy.completePrepareApplicationSinglePolicyType({
+          referenceNumber,
+          exporterHasTradedWithBuyer: true,
+          fullyPopulatedBuyerTradingHistory: true,
+        });
 
-        cy.completeAndSubmitCompanyOrOrganisationForm({});
-        cy.completeAndSubmitConnectionWithTheBuyerForm({});
-        cy.completeAndSubmitTradedWithBuyerForm({ exporterHasTradedWithBuyer: true });
-        cy.completeAndSubmitBuyerFinancialInformationForm({});
+        task.link().click();
 
-        url = `${baseUrl}${ROOT}/${referenceNumber}${CHECK_YOUR_ANSWERS}`;
+        // To get past "Your business" check your answers page
+        cy.completeAndSubmitMultipleCheckYourAnswers({ count: 1 });
+
+        url = `${baseUrl}${ROOT}/${referenceNumber}${YOUR_BUYER}`;
         tradingHistoryUrl = `${baseUrl}${ROOT}/${referenceNumber}${TRADING_HISTORY}`;
+
+        cy.assertUrl(url);
       });
     });
 
     beforeEach(() => {
       cy.saveSession();
+
+      cy.navigateToUrl(url);
     });
 
     after(() => {
@@ -46,12 +60,12 @@ context(
     });
 
     describe('when clicking the `change` link', () => {
-      it(`should redirect to ${TRADING_HISTORY_CHANGE}`, () => {
+      it(`should redirect to ${TRADING_HISTORY_CHECK_AND_CHANGE}`, () => {
         cy.navigateToUrl(url);
 
         summaryList.field(fieldId).changeLink().click();
 
-        cy.assertChangeAnswersPageUrl({ referenceNumber, route: TRADING_HISTORY_CHANGE, fieldId });
+        cy.assertChangeAnswersPageUrl({ referenceNumber, route: TRADING_HISTORY_CHECK_AND_CHANGE, fieldId });
       });
     });
 
@@ -64,14 +78,18 @@ context(
         cy.completeAndSubmitTradingHistoryWithBuyerForm({ outstandingPayments: false });
       });
 
-      it(`should redirect to ${CHECK_YOUR_ANSWERS}`, () => {
-        cy.assertChangeAnswersPageUrl({ referenceNumber, route: CHECK_YOUR_ANSWERS, fieldId });
+      it(`should redirect to ${YOUR_BUYER}`, () => {
+        cy.assertChangeAnswersPageUrl({ referenceNumber, route: YOUR_BUYER, fieldId });
       });
 
       it(`should render the new answer with no ${TOTAL_AMOUNT_OVERDUE} or ${TOTAL_OUTSTANDING_PAYMENTS} rows`, () => {
         cy.assertSummaryListRowValue(summaryList, fieldId, FIELD_VALUES.NO);
         cy.assertSummaryListRowDoesNotExist(summaryList, TOTAL_AMOUNT_OVERDUE);
         cy.assertSummaryListRowDoesNotExist(summaryList, TOTAL_OUTSTANDING_PAYMENTS);
+      });
+
+      it('should retain a `completed` status tag', () => {
+        cy.checkTaskStatusCompleted(status);
       });
 
       describe(`when going back to ${TRADING_HISTORY}`, () => {
