+++ e2e-tests/insurance/cypress/e2e/journeys/check-your-answers/your-buyer/change-your-answers/trading-history/change-your-answers-change-trading-history-yes-to-no.spec.js
@@ -1,21 +1,27 @@
-import { summaryList } from '../../../../../../pages/shared';
-import { FIELD_VALUES } from '../../../../../../constants';
-import { YOUR_BUYER as FIELD_IDS } from '../../../../../../constants/field-ids/insurance/your-buyer';
-import { INSURANCE_ROUTES } from '../../../../../../constants/routes/insurance';
+import { status, summaryList } from '../../../../../../../../pages/shared';
+import partials from '../../../../../../../../partials';
+import { FIELD_VALUES } from '../../../../../../../../constants';
+import { YOUR_BUYER as FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance/your-buyer';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
 
 const { TRADED_WITH_BUYER, OUTSTANDING_PAYMENTS, TOTAL_AMOUNT_OVERDUE, TOTAL_OUTSTANDING_PAYMENTS, FAILED_PAYMENTS } = FIELD_IDS;
 
 const {
   ROOT,
-  YOUR_BUYER: { CHECK_YOUR_ANSWERS, TRADING_HISTORY },
+  CHECK_YOUR_ANSWERS: { YOUR_BUYER },
+  YOUR_BUYER: { TRADING_HISTORY },
 } = INSURANCE_ROUTES;
+
+const { taskList } = partials.insurancePartials;
+
+const task = taskList.submitApplication.tasks.checkAnswers;
 
 const fieldId = TRADED_WITH_BUYER;
 
 const baseUrl = Cypress.config('baseUrl');
 
 context(
-  `Insurance - Your buyer - Change your answers - Trading history - ${TRADED_WITH_BUYER} - Yes to no - As an exporter, I want to change my answers to the trading history section from yes to no`,
+  `Insurance - Check your answers - Your buyer - Trading history - ${TRADED_WITH_BUYER} - Yes to no - As an exporter, I want to change my answers to the trading history section`,
   () => {
     let referenceNumber;
     let url;
@@ -24,15 +30,20 @@ context(
       cy.completeSignInAndGoToApplication({}).then(({ referenceNumber: refNumber }) => {
         referenceNumber = refNumber;
 
-        cy.startInsuranceYourBuyerSection({});
+        cy.completePrepareApplicationSinglePolicyType({
+          referenceNumber,
+          exporterHasTradedWithBuyer: true,
+          fullyPopulatedBuyerTradingHistory: true,
+        });
 
-        cy.completeAndSubmitCompanyOrOrganisationForm({});
-        cy.completeAndSubmitConnectionWithTheBuyerForm({});
-        cy.completeAndSubmitTradedWithBuyerForm({ exporterHasTradedWithBuyer: true });
-        cy.completeAndSubmitTradingHistoryWithBuyerForm({ outstandingPayments: true, failedToPay: true });
-        cy.completeAndSubmitBuyerFinancialInformationForm({});
+        task.link().click();
 
-        url = `${baseUrl}${ROOT}/${referenceNumber}${CHECK_YOUR_ANSWERS}`;
+        // To get past "Your business" check your answers page
+        cy.completeAndSubmitMultipleCheckYourAnswers({ count: 1 });
+
+        url = `${baseUrl}${ROOT}/${referenceNumber}${YOUR_BUYER}`;
+
+        cy.assertUrl(url);
       });
     });
 
@@ -50,8 +61,8 @@ context(
       cy.deleteApplication(referenceNumber);
     });
 
-    it(`should redirect to ${CHECK_YOUR_ANSWERS}`, () => {
-      cy.assertChangeAnswersPageUrl({ referenceNumber, route: CHECK_YOUR_ANSWERS, fieldId });
+    it(`should redirect to ${YOUR_BUYER}`, () => {
+      cy.assertChangeAnswersPageUrl({ referenceNumber, route: YOUR_BUYER, fieldId });
     });
 
     it('should render the new answer, with no other trading history fields', () => {
@@ -60,6 +71,10 @@ context(
       cy.assertSummaryListRowDoesNotExist(summaryList, FAILED_PAYMENTS);
       cy.assertSummaryListRowDoesNotExist(summaryList, TOTAL_AMOUNT_OVERDUE);
       cy.assertSummaryListRowDoesNotExist(summaryList, OUTSTANDING_PAYMENTS);
+    });
+
+    it('should retain a `completed` status tag', () => {
+      cy.checkTaskStatusCompleted(status);
     });
 
     describe(`when changing the answer again from no to yes and going back to ${TRADING_HISTORY}`, () => {
