+++ e2e-tests/insurance/cypress/e2e/journeys/policy/change-your-answers/loss-payee/financial-details/change-your-answers-change-loss-payee-financial-details-uk-to-international.spec.js
@@ -1,4 +1,3 @@
-import partials from '../../../../../../../../partials';
 import { field, summaryList } from '../../../../../../../../pages/shared';
 import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance/policy';
 import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
@@ -13,18 +12,13 @@ const {
 
 const {
   ROOT,
-  POLICY: { LOSS_PAYEE_DETAILS_CHECK_AND_CHANGE, LOSS_PAYEE_FINANCIAL_DETAILS_UK, LOSS_PAYEE_FINANCIAL_DETAILS_INTERNATIONAL_CHECK_AND_CHANGE },
-  CHECK_YOUR_ANSWERS: { TYPE_OF_POLICY },
+  POLICY: { LOSS_PAYEE_DETAILS_CHANGE, LOSS_PAYEE_FINANCIAL_DETAILS_UK, LOSS_PAYEE_FINANCIAL_DETAILS_INTERNATIONAL_CHANGE, CHECK_YOUR_ANSWERS },
 } = INSURANCE_ROUTES;
-
-const { taskList } = partials.insurancePartials;
-
-const task = taskList.submitApplication.tasks.checkAnswers;
 
 const baseUrl = Cypress.config('baseUrl');
 
 context(
-  'Insurance - Change your answers - Policy - Loss payee details - Financial details - UK to International - As an exporter, I want to change my answers to the loss payee section',
+  'Insurance - Policy - Change your answers - Loss payee details - Financial details - UK to International - As an exporter, I want to change my answers to the loss payee section',
   () => {
     let referenceNumber;
     let checkYourAnswersUrl;
@@ -35,22 +29,14 @@ context(
       cy.completeSignInAndGoToApplication({}).then(({ referenceNumber: refNumber }) => {
         referenceNumber = refNumber;
 
-        cy.completePrepareApplicationMultiplePolicyType({
-          referenceNumber,
+        cy.completePolicySection({
           isAppointingLossPayee: true,
           lossPayeeIsLocatedInUK: true,
         });
 
-        task.link().click();
-
-        // To get past previous "Check your answers" pages
-        cy.completeAndSubmitMultipleCheckYourAnswers({ count: 2 });
-
-        checkYourAnswersUrl = `${baseUrl}${ROOT}/${referenceNumber}${TYPE_OF_POLICY}`;
-        lossPayeeDetailsUrl = `${baseUrl}${ROOT}/${referenceNumber}${LOSS_PAYEE_DETAILS_CHECK_AND_CHANGE}`;
-        lossPayeeFinancialInternationalUrl = `${baseUrl}${ROOT}/${referenceNumber}${LOSS_PAYEE_FINANCIAL_DETAILS_INTERNATIONAL_CHECK_AND_CHANGE}`;
-
-        cy.assertUrl(checkYourAnswersUrl);
+        checkYourAnswersUrl = `${baseUrl}${ROOT}/${referenceNumber}${CHECK_YOUR_ANSWERS}`;
+        lossPayeeDetailsUrl = `${baseUrl}${ROOT}/${referenceNumber}${LOSS_PAYEE_DETAILS_CHANGE}`;
+        lossPayeeFinancialInternationalUrl = `${baseUrl}${ROOT}/${referenceNumber}${LOSS_PAYEE_FINANCIAL_DETAILS_INTERNATIONAL_CHANGE}`;
       });
     });
 
@@ -63,12 +49,12 @@ context(
     });
 
     describe(`when clicking the ${NAME} 'change' link`, () => {
-      it(`should redirect to ${LOSS_PAYEE_DETAILS_CHECK_AND_CHANGE}`, () => {
+      it(`should redirect to ${LOSS_PAYEE_DETAILS_CHANGE}`, () => {
         cy.navigateToUrl(checkYourAnswersUrl);
 
         summaryList.field(NAME).changeLink().click();
 
-        cy.assertChangeAnswersPageUrl({ referenceNumber, route: LOSS_PAYEE_DETAILS_CHECK_AND_CHANGE, fieldId: NAME });
+        cy.assertChangeAnswersPageUrl({ referenceNumber, route: LOSS_PAYEE_DETAILS_CHANGE, fieldId: NAME });
       });
     });
 
@@ -77,7 +63,7 @@ context(
         cy.navigateToUrl(checkYourAnswersUrl);
       });
 
-      it(`should redirect to ${LOSS_PAYEE_FINANCIAL_DETAILS_INTERNATIONAL_CHECK_AND_CHANGE} and then ${TYPE_OF_POLICY} after completing (now required) loss payee financial international fields`, () => {
+      it(`should redirect to ${LOSS_PAYEE_FINANCIAL_DETAILS_INTERNATIONAL_CHANGE} and then ${CHECK_YOUR_ANSWERS} after completing (now required) loss payee financial international fields`, () => {
         summaryList.field(NAME).changeLink().click();
 
         cy.assertUrl(`${lossPayeeDetailsUrl}#${NAME}-label`);
@@ -88,7 +74,7 @@ context(
 
         cy.completeAndSubmitLossPayeeFinancialDetailsInternationalForm({});
 
-        cy.assertChangeAnswersPageUrl({ referenceNumber, route: TYPE_OF_POLICY, fieldId: NAME });
+        cy.assertChangeAnswersPageUrl({ referenceNumber, route: CHECK_YOUR_ANSWERS, fieldId: NAME });
       });
 
       it('should render new answers and change links for International financial details, no UK financial details', () => {
