+++ e2e-tests/insurance/cypress/e2e/journeys/your-buyer/change-your-answers/trading-history/change-your-answers-change-trading-history-no-to-yes.spec.js
@@ -1,7 +1,7 @@
-import { YOUR_BUYER as FIELD_IDS } from '../../../../../../constants/field-ids/insurance/your-buyer';
-import { INSURANCE_ROUTES } from '../../../../../../constants/routes/insurance';
-import { summaryList } from '../../../../../../pages/shared';
-import checkSummaryList from '../../../../../../commands/insurance/check-your-buyer-summary-list';
+import { YOUR_BUYER as FIELD_IDS } from '../../../../../../../constants/field-ids/insurance/your-buyer';
+import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
+import { summaryList } from '../../../../../../../pages/shared';
+import checkSummaryList from '../../../../../../../commands/insurance/check-your-buyer-summary-list';
 
 const { TRADED_WITH_BUYER, OUTSTANDING_PAYMENTS, FAILED_PAYMENTS, HAS_PREVIOUS_CREDIT_INSURANCE_COVER_WITH_BUYER, PREVIOUS_CREDIT_INSURANCE_COVER_WITH_BUYER } =
   FIELD_IDS;
